+++ packages/insomnia/src/scripting/run-script-quickjs.ts
@@ -1,204 +1,72 @@
-import type { QuickJSContext, QuickJSHandle } from 'quickjs-emscripten';
-
-import { Console } from '../../../insomnia-scripting-environment/src/objects/console';
 import type { RequestContext } from '../../../insomnia-scripting-environment/src/objects/interfaces';
-import { getQuickJSModule } from '../templating/sandbox/quickjs-runtime';
+import type { WorkerResponse } from './quickjs-script.worker';
 
 /**
- * PoC/experimental alternative to `run-script.ts`'s hidden-`BrowserWindow` execution path. Runs
- * pre-request/after-response scripts inside a QuickJS-WASM VM instead. Only a minimal API surface
- * is bridged in — enough to demonstrate the architecture, not enough to replace the hidden-window
- * sandbox yet. Gated behind `settings.useQuickJsScriptSandbox` (default off).
- *
- * Not supported (throws inside the script if called): insomnia.sendRequest(), insomnia.test()/
- * pm.test(), collectionVariables, vault, cookies, client certificates, and mutating the request.
+ * Client side of the QuickJS-sandbox worker boundary. Runs a script by posting it to a dedicated
+ * Web Worker (`quickjs-script.worker.ts`) instead of calling the engine (`quickjs-script-engine.ts`)
+ * directly on the renderer's main thread — a runaway script blocks only that disposable worker, the
+ * app stays responsive. The worker is created lazily on first use, since most users never enable
+ * `settings.useQuickJsScriptSandbox` and shouldn't pay for a QuickJS-WASM load they never trigger.
  */
-export const runScriptInQuickJs = async ({
+
+let worker: Worker | null = null;
+const pending = new Map<string, { resolve: (value: RequestContext) => void; reject: (err: Error) => void }>();
+
+const getWorker = (): Worker => {
+  if (worker) {
+    return worker;
+  }
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment -- see templating-handler.ts's identical pattern
+  // @ts-ignore -- inso transpiles to commonjs so doesn't play nice with this
+  const instance = new Worker(new URL('quickjs-script.worker.ts', import.meta.url), { type: 'module' });
+
+  instance.addEventListener('message', (event: MessageEvent<WorkerResponse>) => {
+    const { id } = event.data;
+    const request = pending.get(id);
+    if (!request) {
+      return;
+    }
+    pending.delete(id);
+    if ('error' in event.data) {
+      const err = new Error(event.data.error.message);
+      if (event.data.error.name) {
+        err.name = event.data.error.name;
+      }
+      if (event.data.error.stack) {
+        err.stack = event.data.error.stack;
+      }
+      request.reject(err);
+    } else {
+      request.resolve(event.data.result);
+    }
+  });
+
+  // A crash in the worker's global scope (e.g. an unhandled rejection outside our own try/catch)
+  // otherwise leaves every in-flight caller hanging forever. Reject them and drop the worker so the
+  // next call gets a fresh one instead of reusing a dead thread.
+  instance.addEventListener('error', event => {
+    const err = new Error(`QuickJS sandbox worker crashed: ${event.message}`);
+    pending.forEach(request => request.reject(err));
+    pending.clear();
+    if (worker === instance) {
+      worker = null;
+    }
+  });
+
+  worker = instance;
+  return instance;
+};
+
+export const runScriptInQuickJs = ({
   script,
   context,
 }: {
   script: string;
   context: RequestContext;
 }): Promise<RequestContext> => {
-  const QuickJS = await getQuickJSModule();
-  const vm = QuickJS.newContext();
-  const scriptConsole = new Console();
-  const timeoutMs = context.settings.timeout && context.settings.timeout > 0 ? context.settings.timeout : 30_000;
-  const deadline = Date.now() + timeoutMs;
-
-  // Copied in up front and mutated in place by the environment/variables bridges below; whatever
-  // is left in these objects when the script finishes becomes the updated context.
-  const environmentData: Record<string, unknown> = { ...context.environment.data };
-  const variablesData: Record<string, unknown> = { ...context.transientVariables?.data };
-
-  try {
-    // Polled during synchronous execution so a tight sync loop in the script can't bypass the timeout.
-    vm.runtime.setInterruptHandler(() => Date.now() > deadline);
-    vm.runtime.setMemoryLimit(32 * 1024 * 1024);
-
-    installConsole(vm, scriptConsole);
-    installKeyValueBridge(vm, '__envGet', '__envSet', environmentData);
-    installKeyValueBridge(vm, '__varGet', '__varSet', variablesData);
-    setGlobalString(vm, '__requestJSON', JSON.stringify(context.request ?? {}));
-
-    evalOrThrow(vm, BOOTSTRAP, '<quickjs-script-bootstrap>');
-    evalOrThrow(vm, wrapUserScript(script), '<user-script>');
-
-    await driveTaskToCompletion(vm, deadline, timeoutMs);
-  } finally {
-    vm.dispose();
-  }
-
-  return {
-    ...context,
-    environment: {
-      id: context.environment.id,
-      name: context.environment.name,
-      data: environmentData,
-    },
-    transientVariables: {
-      name: context.transientVariables?.name || 'transientVariables',
-      data: variablesData,
-    },
-    logs: scriptConsole.dumpLogsAsArray(),
-  };
-};
-
-const unsupportedApiMessage = (name: string): string =>
-  `${name} is not supported yet by the QuickJS sandbox (proof of concept). Disable "Use QuickJS sandbox for scripts" in Settings > Scripting to use it.`;
-
-const BOOTSTRAP = `
-globalThis.insomnia = {
-  environment: {
-    get: (key) => JSON.parse(__envGet(key)),
-    set: (key, value) => { __envSet(key, JSON.stringify(value === undefined ? null : value)); },
-  },
-  variables: {
-    get: (key) => JSON.parse(__varGet(key)),
-    set: (key, value) => { __varSet(key, JSON.stringify(value === undefined ? null : value)); },
-  },
-  request: (function deepFreeze(value) {
-    if (value && typeof value === 'object') {
-      Object.values(value).forEach(deepFreeze);
-      return Object.freeze(value);
-    }
-    return value;
-  })(JSON.parse(__requestJSON)),
-  sendRequest: () => { throw new Error(${JSON.stringify(unsupportedApiMessage('insomnia.sendRequest()'))}); },
-  test: () => { throw new Error(${JSON.stringify(unsupportedApiMessage('insomnia.test()/pm.test()'))}); },
-};
-globalThis.$ = globalThis.insomnia;
-`;
-
-const wrapUserScript = (script: string): string => `
-globalThis.__task = (async () => {
-  ${script}
-})();
-`;
-
-const installConsole = (vm: QuickJSContext, scriptConsole: Console): void => {
-  const consoleHandle = vm.newObject();
-  (['log', 'warn', 'error', 'info', 'debug'] as const).forEach(level => {
-    const fn = vm.newFunction(level, (...argHandles: QuickJSHandle[]) => {
-      const values = argHandles.map(handle => vm.dump(handle));
-      scriptConsole[level](...values);
-    });
-    vm.setProp(consoleHandle, level, fn);
-    fn.dispose();
+  const id = `quickjs-${Date.now()}-${Math.random().toString(36).slice(2)}`;
+  return new Promise<RequestContext>((resolve, reject) => {
+    pending.set(id, { resolve, reject });
+    getWorker().postMessage({ id, script, context });
   });
-  vm.setProp(vm.global, 'console', consoleHandle);
-  consoleHandle.dispose();
-};
-
-/** Registers `<getName>(key)`/`<setName>(key, jsonValue)` host functions backed by a plain object. */
-const installKeyValueBridge = (
-  vm: QuickJSContext,
-  getName: string,
-  setName: string,
-  store: Record<string, unknown>,
-): void => {
-  const getFn = vm.newFunction(getName, keyHandle => {
-    const key = vm.getString(keyHandle);
-    const value = Object.prototype.hasOwnProperty.call(store, key) ? store[key] : null;
-    return vm.newString(JSON.stringify(value));
-  });
-  setGlobal(vm, getName, getFn);
-
-  const setFn = vm.newFunction(setName, (keyHandle, valueHandle) => {
-    const key = vm.getString(keyHandle);
-    const rawValue = vm.getString(valueHandle);
-    try {
-      store[key] = JSON.parse(rawValue);
-    } catch {
-      store[key] = rawValue;
-    }
-  });
-  setGlobal(vm, setName, setFn);
-};
-
-const setGlobal = (vm: QuickJSContext, name: string, handle: QuickJSHandle): void => {
-  // setProp does NOT take ownership of handle — it must still be disposed here, or the VM leaks
-  // it and asserts ("gc_obj_list" non-empty) when the runtime is later disposed.
-  vm.setProp(vm.global, name, handle);
-  handle.dispose();
-};
-
-const setGlobalString = (vm: QuickJSContext, name: string, value: string): void => {
-  const handle = vm.newString(value);
-  setGlobal(vm, name, handle);
-};
-
-const evalOrThrow = (vm: QuickJSContext, code: string, filename: string): void => {
-  const result = vm.evalCode(code, filename);
-  if (result.error) {
-    const errData = vm.dump(result.error);
-    result.error.dispose();
-    throw toError(errData);
-  }
-  result.value.dispose();
-};
-
-/** Drives `globalThis.__task` (a VM promise wrapping the user script) to completion. */
-const driveTaskToCompletion = async (vm: QuickJSContext, deadline: number, timeoutMs: number): Promise<void> => {
-  const taskHandle = vm.getProp(vm.global, '__task');
-  const resultPromise = vm.resolvePromise(taskHandle);
-  taskHandle.dispose();
-
-  let settled: Awaited<typeof resultPromise> | undefined;
-  resultPromise.then(result => {
-    settled = result;
-  });
-
-  while (!settled) {
-    vm.runtime.executePendingJobs();
-    if (settled) {
-      break;
-    }
-    if (Date.now() > deadline) {
-      throw new Error(`Executing script timeout: ${timeoutMs}`);
-    }
-    await new Promise(resolve => setTimeout(resolve, 0));
-  }
-
-  if (settled.error) {
-    const errData = vm.dump(settled.error);
-    settled.error.dispose();
-    throw toError(errData);
-  }
-  settled.value.dispose();
-};
-
-const toError = (data: unknown): Error => {
-  if (data && typeof data === 'object' && 'message' in data) {
-    const { message, name, stack } = data as { message?: string; name?: string; stack?: string };
-    const err = new Error(message ?? 'QuickJS sandbox error');
-    if (name) {
-      err.name = name;
-    }
-    if (stack) {
-      err.stack = stack;
-    }
-    return err;
-  }
-  return new Error(typeof data === 'string' ? data : JSON.stringify(data));
 };
